+++ src/index.ts
@@ -4,6 +4,5 @@
 
 export { useOnOutsideClick } from './hooks/useOnOutsideClick';
 export { useFocus } from './hooks/useFocus';
-export { arraysIsEqual } from './utils/arrayUtils';
-export { objectIsEmpty, objectsIsEqual } from './utils/objectUtils';
+export { objectIsEmpty, objectsIsEqual, arraysIsEqual } from './utils/objectUtils';
 export { fireAndForget } from './utils/asyncUtils';
